Switch more back to "we", also use "the user" and reword when possible
Keeping "you" when the text talks about something that the reader might
encounter in the future, or when they need to actively do something.
Talk about "the user" as "the player of the game", to distinguish from
the reader (they're probably the same in this case, but I think it's
less confusing this way). Use "we" when introducing new parts of the
code that we're working on together with the reader. Some spots didn't
really need to make this decision; I reworded them to be more generic.