re-work last few chapters
Now that "Advanced type systems" is pretty much only two things, merging
the lifetimes chapter, which is very short, into it makes a lot of
sense.

In addition, we kinda wanted to see if we could fit another project in
here anyway, so let's do this and then it makes room!