add fuzzing to Listing 16-1, 2
For a reader that will try to run the code in Listings 16-1 and 16-2, she might find the output fundamentally different from the one showed here. And the advice to raise the numbers, while valid, she might need to pick a number as larger as 5000, just to see this happens once.