Tweak description of ? operator
Amends the description of the ? operator's behaviour: if Ok, returns the value inside the Ok; if Err, returns the whole Err, not just the value inside.

Clarifies the para following.  'used with' is imprecise.  The user 'uses' the ? operator with error values generally, but it's not clear whether the error value parameter or the returned error value of ? is the subject of the sentence.  Operators do not generally 'use' parameters.  More precisely, the error value 'taken by' ? goes through `from` before being returned.