Remove section on the anonymous lifetime; this doesn't make sense
I think this section needs more motivation or else it doesn't belong in
the book and should be documented somewhere else. It also appears,
according to the tracking issue, that there is more to be done, which I
think will affect best practices and how this is used in the wild.

Basically, I don't think this feature is ready for the book yet.

Reopens #1273