Correct chapter 20 final listing
In chapter 20 on writing a web server, one of the listings includes `.take(2)` on the TCP stream iterator to only handle 2 requests and then quit. This is as an example to demonstrate that the web server shuts down gracefully. However, this has slipped into the final code listing, which I don't think should include this arbitrary restriction. This pull request amends this error.