Disable playground button on listings in ch 12 that use CLI args
You can't pass CLI args to the playground, so it doesn't make sense to
run them in the browser. Also, these were marked as should_panic,
because if you *do* run them in the browser or without arguments, they
panic, but that's not the happy path we're writing, so "should" panic
isn't really correct.

Fixes #2370.