Regenerate ch09-02 error messages
In 370dbc063 the advantages of `expect` have been clarified to remove
the statements that it helps with locating the error. This is accurate
since 1.46.0 introduced the `#[track_caller]` attribute.

The original patch however did not regenerate all the errors with the
latest compiler, so this patch updates the errors to avoid potential
confusion.