Explicitly state "compilation modes"
I'm a Rust newbie. When I first read the line "Wrap in all modes...", I first thought that "mode" referred to different modes of the `wrapping_` method. It took me a minute to realise that "mode" referred to the _compilation_ mode! (Maybe I'm a bit sleepy today!) So I'd propose that the text explicitly says "compilation modes", to make the meaning as clear as possible :)