Fix the expected substring in ch11-01
According to Listing 11-9, the substring expected to be included is
"less than or equal to 100" (not "Guess value must be less than or equal
to 100" as written in the book, which is strictly speaking not the
expected substring - it's almost the entire panic message except the
final ", got {}." part). See
https://github.com/rust-lang/book/blob/f78b8a190adcdc459241d05b6a8052410619ba5c/listings/ch11-writing-automated-tests/listing-11-09/src/lib.rs#L31:

```rust
    #[should_panic(expected = "less than or equal to 100")]
```

Also I think this was the only place in this chapter / in the book where
I saw a string quotation in inline code additionally surrounded with
single quotes `'`, so I removed them.