Ch. 17: build a `retry` implementation
This is good to know about, but has significant problems, mostly down to
the issues Niko covers in [this post][post] regarding the way ownership
and closures interact. See [this Stack Overflow answer][so], which
applies that dynamic specifically to a `retry` implementation which
happens to be more or less identical to what I built here!

[post]: https://smallcultfollowing.com/babysteps/blog/2023/05/09/giving-lending-and-async-closures/
[answer]: https://stackoverflow.com/questions/75533630/how-to-write-a-retry-function-in-rust-that-involves-async