Ch. 17: Remove TODO about `Stream` stabilization
Per the [async working group][wg], the _status quo_ for this is unlikely
to change before we go to print. We can of course update the online copy
as soon as it *does* change, and if it happens to change such that we
can sneak it in before going to press, great.

[wg]: https://rust-lang.zulipchat.com/#narrow/stream/187312-wg-async/topic/Rename.20AsyncIterator.20back.20to.20Stream.2C.20introduce.20an.20AFIT-based/near/475300446