Add HTML-based redirects for previous chapter numbering
Unlike the redirects we have in place for the previous editions, this
simply performs an in-place redirect, using the built-in mechanism for
this in mdbook. As an alternative, we could hand-craft HTML pages to do
what the redirect pages for the first and second edition do: present a
page and let the user click through proactively.