infra: create mdbook-trpl-figure preprocessor
Building on the shared infra from `mdbook_trpl`, this preprocessor takes
nicely accessible and semantic HTML input like this:

    <figure>
    
    <img src="https://www.example.com/something.jpg">
    
    <figcaption>Figure 12-34: An illustration of something</figcaption>
    
    </figure>
    
It produces output like this:

    <img src="https://www.example.com/something.jpg">
    
    Figure 12-34: An illustration of something

This matches what we need for the nostarch output. Accordingly, wire up
the `nostarch/book.toml` to use this. There is no need to worry about
ordering of the two preprocessors, because in `Simple` mode the listing
preprocessor emits plain text output for listings (much as this does for
figures).