Crarified the misunderstanding b/w crates, module, items
The issue is while in std::io::stdin:  
std is the standard library in Rust.
io is a module within the std library.
stdin is a function within the io module, but the book said "io library" it does confused too much as a newbie the user won't impose his thoughts for a moment he/she will rather trust the BOOK that ::io is indeed a library (with confusion in his mind). so it needs to be fixed.