Ferris: simplify line counting for size determination
This is a trivial thing, but I found the previous “replace a terminal
newline and then count the rest” rather less than obvious. Make the
logic more explicit and, as a fun bonus, skip allocating an extra string
along the way. (Not that it matters, in this case, but it’s always
nice to make that kind of improvement.)